# contracts/openapi/yudha-api.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "YUDHA Public API",
    "version": "1.0.0",
    "description": "PRD-aligned public contract. x-delivery-gate identifies the acceptance gate that activates conformance enforcement."
  },
  "servers": [{ "url": "http://localhost:3000" }],
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/economy": { "get": { "x-delivery-gate": 3, "responses": { "200": { "description": "Refreshed economy state", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DataEnvelope" } } } }, "401": { "$ref": "#/components/responses/Error" } } } },
    "/economy/catalog": { "get": { "x-delivery-gate": 3, "responses": { "200": { "description": "Economy catalog", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DataEnvelope" } } } }, "401": { "$ref": "#/components/responses/Error" } } } },
    "/economy/energy-purchases": { "post": { "x-delivery-gate": 3, "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PurchaseEnergyRequest" } } } }, "responses": { "200": { "$ref": "#/components/responses/Data" }, "409": { "$ref": "#/components/responses/Error" } } } },
    "/economy/ad-rewards/claims": { "post": { "x-delivery-gate": 3, "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ClaimAdRewardRequest" } } } }, "responses": { "200": { "$ref": "#/components/responses/Data" }, "403": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" } } } },
...
  },
  "components": {
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" } },
...
    "responses": {
      "Data": { "description": "Successful response", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DataEnvelope" } } } },
...
      "Error": { "description": "Canonical error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } }
    },
    "schemas": {
...
      "IdempotencyKey": { "type": "string", "minLength": 1, "maxLength": 160 },
...
      "DataEnvelope": { "type": "object", "required": ["data"], "properties": { "data": {} } },
...
      "ErrorEnvelope": { "type": "object", "required": ["error"], "properties": { "error": { "type": "object", "required": ["code", "message", "details", "requestId"], "properties": { "code": { "type": "string", "enum": ["AUTH_REQUIRED", "FORBIDDEN", "VALIDATION_FAILED", "NOT_FOUND", "CONFLICT", "IDEMPOTENCY_KEY_REUSED", "INSUFFICIENT_Y_COIN", "INSUFFICIENT_ENERGY", "ENERGY_CAP_REACHED", "AD_REWARD_LIMIT_REACHED", "AD_VERIFICATION_FAILED", "PRO_ALREADY_ACTIVE", "PRO_SKIN_NOT_ELIGIBLE", "FEATURE_DISABLED", "ENTITLEMENT_REQUIRED", "QUEUE_UNAVAILABLE", "ROOM_CODE_INVALID", "ACTION_REJECTED", "PROVIDER_UNAVAILABLE", "RATE_LIMITED", "INTERNAL_ERROR"] }, "message": { "type": "string" }, "details": { "type": "object" }, "requestId": { "type": "string" } } } } },
...
      "PurchaseEnergyRequest": { "type": "object", "required": ["idempotencyKey", "packageId"], "properties": { "idempotencyKey": { "$ref": "#/components/schemas/IdempotencyKey" }, "packageId": { "type": "string" } } },
      "ClaimAdRewardRequest": { "type": "object", "required": ["idempotencyKey", "rewardType", "placementId", "proofToken"], "properties": { "idempotencyKey": { "$ref": "#/components/schemas/IdempotencyKey" }, "rewardType": { "type": "string", "enum": ["energy", "y_coin"] }, "placementId": { "type": "string" }, "proofToken": { "type": "string" } } },
...
    }
  }
}